requirements.txt
# Core LLM & Agent Framework
langchain>=0.1.0
langchain-groq>=0.1.0
langchain-community>=0.0.20

# Graph Database
neo4j>=5.0.0

# Data Validation
pydantic>=2.0.0

# UI
streamlit>=1.30.0

# Environment & Config
python-dotenv>=1.0.0

# Utilities
tenacity>=8.0.0

# Visualization
pyvis>=0.3.0

# Testing
pytest>=7.0.0
pytest-mock>=3.0.0
